+++ similarity/clustering/mean_shift.h
@@ -5,7 +5,7 @@
 #include "PairWiseCluster.h"
 
 // Function to peform the clustering
-PairWiseClusters * clustering(double *a2, int x, double *b2, int y, int n2,
+PairWiseClusters * clustering(float *a2, int x, float *b2, int y, int n2,
     EMatrix * ematrix, CCMParameters params, float bw, int level);
