+++ similarity/methods/MISimilarity.h
@@ -2,7 +2,7 @@
 #define _MISIMILARITY_
 
 #include <math.h>
-#include <gsl/gsl_statistics.h>
+#include <gsl/gsl_statistics_float.h>
 #include <gsl/gsl_bspline.h>
 #include "PairWiseSimilarity.h"
 
@@ -12,16 +12,16 @@
 class MISimilarity: public PairWiseSimilarity {
 
   private:
-    double mi_bins;
-    double mi_degree;
+    float mi_bins;
+    float mi_degree;
 
-    double calculateBSplineMI(double *v1, double *v2, int n, int m, int k,
-        double xmin, double ymin, double xmax, double ymax);
+    float calculateBSplineMI(float *v1, float *v2, int n, int m, int k,
+        float xmin, float ymin, float xmax, float ymax);
 
   public:
     MISimilarity(PairWiseSet * pws, int min_obs, int * samples);
-    MISimilarity(PairWiseSet * pws, int min_obs, double mi_bins, double mi_degree);
-    MISimilarity(PairWiseSet * pws, int min_obs, int * samples, double mi_bins, double mi_degree);
+    MISimilarity(PairWiseSet * pws, int min_obs, float mi_bins, float mi_degree);
+    MISimilarity(PairWiseSet * pws, int min_obs, int * samples, float mi_bins, float mi_degree);
     ~MISimilarity();
 
     void run();
